replace constructorof with something not UB (#102)
* replace constructorof with something not UB

It is confusing that `constructorof` is documented as returning the constructor of `T`, which is normally defined to be `T`, and then typically does not return `T`, and so does not usually construct `T`. However, this PR is intended just to remove the undefined behavior here of using an unsound generated function (examining mutable state with getfield).

* Update ConstructionBase.jl

* Update ConstructionBase.jl

* Update ConstructionBase.jl

* add pre to CI

* upd ci

* drop support for unmaintained versions of Julia

* delete remaining branching for unsupported versions of Julia

* remove pieces specific to old julia versions

---------

Co-authored-by: Jameson Nash <vtjnash@gmail.com>
Co-authored-by: Alexander Plavin <alexander@plav.in>